Update docs after reading shutil.rmtree documentation
I think this issue is a bit murky. I agree with the comments, which is
to say that the `shutil.rmtree` function should be more robust (and not
allow a 3rd party to try to fix things).

That being said, maybe there's some learning in here that can be applied
to the new `onexc` function.